[wallet] refs #45 - Divide Qt Models into different packages
